sandbox materialeditor tests until the editor_test re-factor is complete
